+++ spec/rails_helper.rb
@@ -39,7 +39,7 @@ RSpec.configure do |config|
   config.before(:each) do
     @signed_in_user = create(:securial_user)
     @signed_in_session = create(:securial_session, user: @signed_in_user)
-    @valid_token = Securial::Sessions::SessionEncoder.encode(@signed_in_session)
+    @valid_token = Securial::Auth::AuthEncoder.encode(@signed_in_session)
     @valid_headers = { "Authorization" => "Bearer #{@valid_token}", "User-Agent" => "Ruby/RSpec" }
     @invalid_headers = { "Authorization" => "Bearer bad_token" }
   end
@@ -63,6 +63,7 @@ RSpec.configure do |config|
         t.string :password_digest
         t.string :reset_password_token
         t.datetime :reset_password_token_created_at
+        t.datetime :password_changed_at
         t.timestamps
       end
     end